+++ app/src/main/AndroidManifest.xml
@@ -15,16 +15,16 @@
         android:theme="@style/AppTheme"
         tools:targetApi="31">
         <activity
-            android:name=".ExportActivity"
+            android:name=".activities.ExportActivity"
             android:exported="false" />
         <activity
-            android:name=".ReportActivity"
+            android:name=".activities.ReportActivity"
             android:exported="false" />
         <activity
-            android:name=".RecordActivity"
+            android:name=".activities.RecordActivity"
             android:exported="false" />
         <activity
-            android:name=".MainActivity"
+            android:name=".activities.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
