+++ app/src/main/java/com/samaya/myspendings/activities/RecordActivity.java
@@ -1,4 +1,4 @@
-package com.samaya.myspendings;
+package com.samaya.myspendings.activities;
 
 import android.content.Intent;
 import android.os.Build;
@@ -16,7 +16,10 @@ import com.google.android.material.datepicker.MaterialDatePicker;
 import com.google.android.material.datepicker.MaterialPickerOnPositiveButtonClickListener;
 import com.google.android.material.textfield.TextInputEditText;
 import com.google.android.material.timepicker.MaterialTimePicker;
+import com.samaya.myspendings.R;
+import com.samaya.myspendings.utils.Utils;
 import com.samaya.myspendings.db.entity.Spendings;
+import com.samaya.myspendings.fragments.SpendingsViewModel;
 
 import java.util.Calendar;
 import java.util.TimeZone;
