+++ app/src/main/java/com/samaya/myspendings/adapters/ReportFragmentAdapter.java
@@ -1,10 +1,12 @@
-package com.samaya.myspendings;
+package com.samaya.myspendings.adapters;
 
 import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentManager;
 import androidx.lifecycle.Lifecycle;
 import androidx.viewpager2.adapter.FragmentStateAdapter;
+
+import com.samaya.myspendings.fragments.ReportFragment;
 
 public class ReportFragmentAdapter  extends FragmentStateAdapter {
     public ReportFragmentAdapter(@NonNull FragmentManager fragmentManager, @NonNull Lifecycle lifecycle) {
