+++ app/src/main/res/layout/activity_record.xml
@@ -6,7 +6,7 @@
     android:orientation="vertical"
     android:paddingTop="5dp"
     android:paddingBottom="5dp"
-    tools:context=".RecordActivity">
+    tools:context=".activities.RecordActivity">
 
     <!-- TODO: Update blank fragment layout -->
