+++ app/src/main/res/layout/activity_report.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ReportActivity">
+    tools:context=".activities.ReportActivity">
 
 
     <com.google.android.material.tabs.TabLayout
