+++ app/src/main/res/layout/fragment_report_all.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ReportFragment">
+    tools:context=".fragments.ReportFragment">
 
         <com.github.mikephil.charting.charts.LineChart
             android:id="@+id/chart_all"
